backend/requirements.txt
# backend/requirements.txt
Django==4.2.15
djangorestframework==3.15.2
djoser==2.2.3
django-filter==24.3
drf-extra-fields==3.7.0
Pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
django-cors-headers==4.5.0
